Move unused function into test files
